refactor: split root.go into check.go, upgrade.go, and root.go
Pure structural refactor — no logic changes. Moves code into three
files by responsibility:

- check.go: validation types, check command, check/validate logic, JSON output
- upgrade.go: upgrade types, upgrade command, upgrade logic, preview/diff helpers
- root.go: wiring (newRootCmd), shared helpers, doctor command

All files remain in package main. Build, vet, and tests pass.
